feat: 로또 파일 위치 변경 및 유효성 검사 기능 구현
